+++ duckscript_sdk/src/sdk/std/debug/dump_instructions/mod_test.rs
@@ -1,0 +1,17 @@
+use super::*;
+use crate::test;
+use crate::test::CommandValidation;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create(""));
+}
+
+#[test]
+fn run_no_args() {
+    test::run_script_and_validate(
+        vec![create("")],
+        "out = dump_instructions",
+        CommandValidation::Contains("out".to_string(), "dump_instructions".to_string()),
+    );
+}